requirements.txt
Django==2.0
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-resized==0.3.9
django-smart-selects==1.5.4
mysqlclient==1.4.6
Pillow==6.2.1
pytz==2019.3
